#212 Arreglando problema para enviar id del grupo a eliminar

+++ web/grails-app/controllers/com/modulus/uno/GroupNotificationController.groovy
@@ -51,7 +51,7 @@ class GroupNotificationController {
     }
 
     def delete(){
-      groupNotificationService.deleteGroup(params.id)
+      groupNotificationService.deleteGroup(params.id.toLong())
       redirect action:"index", method:"GET"
     }
